Remove unused uiControls parameter from RendererManager
